refactor: professional API polish and convenience helpers
- Change Service.Port from int to uint16 (matches DNS wire format)
- Add ParseTXT() helper for RFC 6763 §6 key=value TXT parsing
- Add DiscoverServices() convenience method (full DNS-SD sequence)
- Add ServiceInstance type for resolved service results
- Create responder/doc.go with comprehensive package documentation
- Add prominent blocking warning to Register() godoc
- Replace 3 skipped tests: implement QU bit test, remove stale skips
- Add ParseTXT test suite (6 cases)

+++ internal/responder/registry_test.go
@@ -147,7 +147,7 @@ func TestRegistry_ConcurrentAccess(t *testing.T) {
 			service := &Service{
 				InstanceName: formatInstanceName("Service", id),
 				ServiceType:  "_http._tcp.local",
-				Port:         8080 + id,
+				Port:         uint16(8080 + id),
 			}
 
 			err := registry.Register(service)
@@ -191,7 +191,7 @@ func TestRegistry_ConcurrentReadWrite(_ *testing.T) {
 		service := &Service{
 			InstanceName: formatInstanceName("Service", i),
 			ServiceType:  "_http._tcp.local",
-			Port:         8080 + i,
+			Port:         uint16(8080 + i),
 		}
 		_ = registry.Register(service)
 	}
@@ -218,7 +218,7 @@ func TestRegistry_ConcurrentReadWrite(_ *testing.T) {
 			service := &Service{
 				InstanceName: formatInstanceName("Service", id),
 				ServiceType:  "_http._tcp.local",
-				Port:         8080 + id,
+				Port:         uint16(8080 + id),
 			}
 			_ = registry.Register(service)
 		}(i)
